refactor: extract shared convex-geometry primitives into geometryHelper
Move clipPolygonByPlane, VertexPool/coordKey, sortPolygonVertices, the
Newell normal/area helpers, and a signed-distance helper out of
printCutting.ts into geometryHelper.ts so the gap-cutting and print-prep
pipelines share one implementation (resolves the known helper
duplication). printCutting re-exports VertexPool/ClipPlane for existing
test imports. No behavior change.
